add modules, fix exceptions
